[Refactor] artifact-type-definition.cpp/h からartifact-list.cpp/h を分離させた
