+++ VisualStudio/Hengband/Hengband.vcxproj.filters
@@ -2571,6 +2571,9 @@
     <ClCompile Include="..\..\src\system\artifact\artifact-definition.cpp">
       <Filter>system\artifact</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\system\artifact\artifact-list.cpp">
+      <Filter>system\artifact</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\src\combat\shoot.h">
@@ -5581,6 +5584,9 @@
       <Filter>system\floor</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\system\artifact\artifact-definition.h">
+      <Filter>system\artifact</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\system\artifact\artifact-list.h">
       <Filter>system\artifact</Filter>
     </ClInclude>
   </ItemGroup>
