+++ src/Makefile.am
@@ -1256,7 +1256,8 @@ hengband_SOURCES = \
 	system/angband-exceptions.h \
 	system/angband-system.h system/angband-system.cpp \
 	system/angband-version.cpp system/angband-version.h \
-	system/artifact/artifact-type-definition.cpp system/artifact/artifact-type-definition.h \
+	system/artifact/artifact-definition.cpp system/artifact/artifact-definition.h \
+	system/artifact/artifact-list.cpp system/artifact/artifact-list.h \
 	system/building-type-definition.cpp system/building-type-definition.h \
 	system/grid-type-definition.cpp system/grid-type-definition.h \
 	system/h-basic.h system/h-config.h \
