+++ src/birth/game-play-initializer.cpp
@@ -14,6 +14,7 @@
 #include "player/digestion-processor.h"
 #include "player/player-spell-status.h"
 #include "system/artifact/artifact-definition.h"
+#include "system/artifact/artifact-list.h"
 #include "system/baseitem/baseitem-definition.h"
 #include "system/baseitem/baseitem-list.h"
 #include "system/building-type-definition.h"
