+++ src/dungeon/quest.cpp
@@ -13,6 +13,7 @@
 #include "player-status/player-energy.h"
 #include "player/player-status.h"
 #include "system/artifact/artifact-definition.h"
+#include "system/artifact/artifact-list.h"
 #include "system/floor/floor-info.h" // @todo 相互参照、将来的に削除する.
 #include "system/grid-type-definition.h"
 #include "system/item-entity.h"
