+++ src/floor/fixed-map-generator.cpp
@@ -17,6 +17,7 @@
 #include "object-enchant/item-magic-applier.h"
 #include "sv-definition/sv-scroll-types.h"
 #include "system/artifact/artifact-definition.h"
+#include "system/artifact/artifact-list.h"
 #include "system/baseitem/baseitem-definition.h"
 #include "system/baseitem/baseitem-list.h"
 #include "system/dungeon/dungeon-definition.h"
