+++ src/info-reader/dungeon-reader.cpp
@@ -9,6 +9,7 @@
 #include "io/tokenizer.h"
 #include "main/angband-headers.h"
 #include "system/artifact/artifact-definition.h"
+#include "system/artifact/artifact-list.h"
 #include "system/baseitem/baseitem-definition.h"
 #include "system/baseitem/baseitem-list.h"
 #include "system/dungeon/dungeon-definition.h"
