+++ src/load/item/item-loader-base.cpp
@@ -3,6 +3,7 @@
 #include "load/angband-version-comparer.h"
 #include "load/load-util.h"
 #include "system/artifact/artifact-definition.h"
+#include "system/artifact/artifact-list.h"
 #include "system/baseitem/baseitem-definition.h"
 #include "system/baseitem/baseitem-list.h"
 #include "util/bit-flags-calculator.h"
