+++ src/monster-floor/special-death-switcher.cpp
@@ -28,6 +28,7 @@
 #include "sv-definition/sv-protector-types.h"
 #include "sv-definition/sv-weapon-types.h"
 #include "system/artifact/artifact-definition.h"
+#include "system/artifact/artifact-list.h"
 #include "system/baseitem/baseitem-allocation.h"
 #include "system/enums/monrace/monrace-id.h"
 #include "system/floor/floor-info.h"
