+++ src/system/artifact/artifact-definition.h
@@ -53,34 +53,3 @@ private:
     bool evaluate_rarity() const;
     bool evaluate_shallow_baseitem(int making_level) const;
 };
-
-class ItemEntity;
-class ArtifactList : public util::AbstractMapWrapper<FixedArtifactId, ArtifactType> {
-public:
-    ArtifactList(const ArtifactList &) = delete;
-    ArtifactList(ArtifactList &&) = delete;
-    ArtifactList &operator=(const ArtifactList &) = delete;
-    ArtifactList &operator=(ArtifactList &&) = delete;
-    ~ArtifactList() = default;
-
-    static ArtifactList &get_instance();
-    const ArtifactType &get_artifact(const FixedArtifactId fa_id) const;
-    ArtifactType &get_artifact(const FixedArtifactId fa_id);
-
-    bool order(const FixedArtifactId id1, const FixedArtifactId id2) const;
-    void emplace(const FixedArtifactId fa_id, ArtifactType &&artifact);
-    void reset_generated_flags();
-    tl::optional<ItemEntity> try_make_instant_artifact(int making_level) const;
-
-private:
-    ArtifactList() = default;
-    static ArtifactList instance;
-    static ArtifactType dummy;
-
-    std::map<FixedArtifactId, ArtifactType> artifacts{};
-
-    std::map<FixedArtifactId, ArtifactType> &get_inner_container() override
-    {
-        return this->artifacts;
-    }
-};
