+++ src/system/artifact/artifact-list.cpp
@@ -1,0 +1,80 @@
+#include "system/artifact/artifact-list.h"
+#include "artifact/fixed-art-types.h"
+#include "system/artifact/artifact-definition.h"
+#include "system/item-entity.h"
+
+ArtifactList ArtifactList::instance{};
+
+ArtifactType ArtifactList::dummy{};
+
+ArtifactList &ArtifactList::get_instance()
+{
+    return instance;
+}
+
+const ArtifactType &ArtifactList::get_artifact(const FixedArtifactId fa_id) const
+{
+    if (fa_id == FixedArtifactId::NONE) {
+        return dummy;
+    }
+
+    return this->artifacts.at(fa_id);
+}
+
+ArtifactType &ArtifactList::get_artifact(const FixedArtifactId fa_id)
+{
+    if (fa_id == FixedArtifactId::NONE) {
+        return dummy;
+    }
+
+    return this->artifacts.at(fa_id);
+}
+
+bool ArtifactList::order(const FixedArtifactId id1, const FixedArtifactId id2) const
+{
+    const auto &artifact1 = this->get_artifact(id1);
+    const auto &artifact2 = this->get_artifact(id2);
+    if (artifact1.bi_key < artifact2.bi_key) {
+        return true;
+    }
+
+    if (artifact1.bi_key > artifact2.bi_key) {
+        return false;
+    }
+
+    if (artifact1.level < artifact2.level) {
+        return true;
+    }
+
+    if (artifact1.level > artifact2.level) {
+        return false;
+    }
+
+    return id1 < id2;
+}
+
+void ArtifactList::emplace(const FixedArtifactId fa_id, ArtifactType &&artifact)
+{
+    this->artifacts.emplace(fa_id, std::move(artifact));
+}
+
+void ArtifactList::reset_generated_flags()
+{
+    for (auto &[_, artifact] : this->artifacts) {
+        artifact.is_generated = false;
+    }
+}
+
+tl::optional<ItemEntity> ArtifactList::try_make_instant_artifact(int making_level) const
+{
+    for (const auto &[fa_id, artifact] : this->artifacts) {
+        const auto bi_key = artifact.try_make_instant_artifact(making_level);
+        if (bi_key) {
+            ItemEntity instant_artifact(*bi_key);
+            instant_artifact.fa_id = fa_id;
+            return instant_artifact;
+        }
+    }
+
+    return tl::nullopt;
+}
