+++ src/wizard/fixed-artifacts-spoiler.cpp
@@ -2,6 +2,7 @@
 #include "io/files-util.h"
 #include "system/angband-system.h"
 #include "system/artifact/artifact-definition.h"
+#include "system/artifact/artifact-list.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
 #include "util/angband-files.h"
