+++ src/wizard/wizard-item-modifier.cpp
@@ -25,6 +25,7 @@
 #include "spell-kind/spells-perception.h"
 #include "spell/spells-object.h"
 #include "system/artifact/artifact-definition.h"
+#include "system/artifact/artifact-list.h"
 #include "system/baseitem/baseitem-allocation.h"
 #include "system/baseitem/baseitem-definition.h"
 #include "system/baseitem/baseitem-list.h"
