+++ src/wizard/wizard-special-process.cpp
@@ -55,6 +55,7 @@
 #include "spell/spells-status.h"
 #include "status/bad-status-setter.h"
 #include "system/artifact/artifact-definition.h"
+#include "system/artifact/artifact-list.h"
 #include "system/dungeon/dungeon-definition.h"
 #include "system/dungeon/dungeon-list.h"
 #include "system/enums/dungeon/dungeon-id.h"
